requirements.txt
Django~=2.1.3
djangorestframework~=3.9.0
djangorestframework-simplejwt~=3.3
django-cors-headers~=2.4.0
channels~=2.1.5
daphne~=2.2.3
whitenoise~=4.1.2